Move ReadyCheckingSideInputReader to util
This SideInputReader allows callers to check for a side input being
available before attempting to read the contents
